Download Issue Fix

+++ app/src/main/java/com/netcast/radio/download/DownloadUsingMediaStore.kt
@@ -1,6 +1,9 @@
 package com.netcast.radio.download
 
 
+import android.annotation.SuppressLint
+import android.app.Activity
+import android.app.AlertDialog
 import android.app.DownloadManager
 import android.app.Notification
 import android.app.NotificationManager
@@ -12,6 +15,7 @@ import android.os.AsyncTask
 import android.os.Build
 import android.os.Environment
 import android.provider.MediaStore
+import android.widget.Toast
 import androidx.core.app.NotificationCompat
 import com.netcast.radio.base.AppSingelton
 import com.netcast.radio.db.AppDatabase
@@ -28,6 +32,8 @@ class DownloadUsingMediaStore(var data: Data, var context: Context) :
     private var notificationManager: NotificationManager? = null
     private val DOWNLOAD_NOTIFICATION_ID = 1
     var relativePath = ""
+    var downloadId:Long = 0
+    private var downloadManager:DownloadManager?=null
 
     private fun updateNotification(progress: Int) {
         val builder: NotificationCompat.Builder =
@@ -80,15 +86,11 @@ class DownloadUsingMediaStore(var data: Data, var context: Context) :
                     MediaStore.Audio.Media.EXTERNAL_CONTENT_URI,
                     values
                 )
-                audioOutStream = context.contentResolver.openOutputStream(uri!!)!!
-                relativePath = getRealPathFromURI(context, uri)
+                relativePath = getRealPathFromURI(context, uri!!)
             } else {
                 Environment.getExternalStorageDirectory().path + "/${fileName}"
-                val audio = File(
-                    Environment.getExternalStoragePublicDirectory(Environment.DIRECTORY_MUSIC)
-                        .toString(), fileName
-                )
-                audioOutStream = FileOutputStream(audio)
+                val audio = File(Environment.getExternalStoragePublicDirectory(Environment.DIRECTORY_MUSIC).toString(), fileName)
+
                 relativePath = audio.path
             }
             AppSingelton.currentDownloading = data.id.toString()
@@ -182,28 +184,61 @@ class DownloadUsingMediaStore(var data: Data, var context: Context) :
         if (appDatabase == null)
             appDatabase = AppDatabase.getDatabaseClient(context)
         val file1 = File(relativePath)
+        if (downloadManager!=null && isDownloadingInProgress(context)) {
+            (context as Activity).runOnUiThread(Runnable {
+                showAlertDialog("Alert","File Already Downlaoding")
+            })
+        } else {
+            if (file1.exists()) {
+                (context as Activity).runOnUiThread(Runnable {
+                    showAlertDialog("Alert","File Already Exist")
+                })
 
-        if (file1.exists()) {
-            file1.delete()
-        }
-        // fileName -> fileName with extension
-        val request = DownloadManager.Request(Uri.parse(url))
-            .setAllowedNetworkTypes(DownloadManager.Request.NETWORK_WIFI or DownloadManager.Request.NETWORK_MOBILE)
-            .setTitle("Downloading $title")
-            .setDescription(desc)
-            .setNotificationVisibility(DownloadManager.Request.VISIBILITY_VISIBLE_NOTIFY_COMPLETED)
-            .setAllowedOverRoaming(true)
-            .setAllowedOverMetered(true)
-            .setDestinationUri(Uri.fromFile(file1))
+                file1.delete()
+            }
+            // fileName -> fileName with extension
+            val request = DownloadManager.Request(Uri.parse(url))
+                .setAllowedNetworkTypes(DownloadManager.Request.NETWORK_WIFI or DownloadManager.Request.NETWORK_MOBILE)
+                .setTitle("Downloading $title")
+                .setDescription(desc)
+                .setNotificationVisibility(DownloadManager.Request.VISIBILITY_VISIBLE_NOTIFY_COMPLETED)
+                .setAllowedOverRoaming(true)
+                .setAllowedOverMetered(true)
+                .setDestinationUri(Uri.fromFile(file1))
 //            .setDestinationInExternalPublicDir(relativePath, fileName)
-        val downloadManager = context.getSystemService(Context.DOWNLOAD_SERVICE) as DownloadManager
-        val downloadID = downloadManager.enqueue(request)
+            downloadManager = context.getSystemService(Context.DOWNLOAD_SERVICE) as DownloadManager
+
+            downloadId = downloadManager?.enqueue(request)!!
 //        data.fileURI = Environment.DIRECTORY_DOWNLOADS+"/${fileName}"
-        data.fileURI = relativePath
-        CoroutineScope(Dispatchers.IO).launch {
-            appDatabase!!.appDap().insertOfflineEpisode(data)
+            data.fileURI = relativePath
+            CoroutineScope(Dispatchers.IO).launch {
+                appDatabase!!.appDap().insertOfflineEpisode(data)
+            }
+            AppSingelton.currentDownloading = ""
         }
-        AppSingelton.currentDownloading = ""
+
+
     }
+
+    private fun showAlertDialog(title: String, message: String) {
+        val builder: AlertDialog.Builder = AlertDialog.Builder(context)
+        builder.setTitle(title)
+            .setMessage(message)
+            .setPositiveButton("OK", null)
+            .show()
+    }
+    private fun isDownloadingInProgress(context: Context): Boolean {
+//         downloadManager = context.getSystemService(Context.DOWNLOAD_SERVICE) as DownloadManager
+
+        val query = DownloadManager.Query()
+        query.setFilterByStatus(DownloadManager.STATUS_RUNNING)
+
+        val cursor = downloadManager?.query(query)
+        val inProgress = cursor?.count!! > 0
+        cursor?.close()
+
+        return inProgress
+    }
+
 
 }
